(refactor) ui components
